5. Capstone Project/Bookmarkie/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
python-dotenv==0.13.0
pycryptodome==3.9.7
six==1.12.0
python-jose[pycryptodome]
gunicorn==20.0.4